refactor: migrate fade upwards page transition builder to widgets
