+++ packages/flutter/test/material/page_test.dart
@@ -16,76 +16,6 @@ import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   testWidgets(
-    'test page transition (_FadeUpwardsPageTransition)',
-    (WidgetTester tester) async {
-      await tester.pumpWidget(
-        MaterialApp(
-          home: const Material(child: Text('Page 1')),
-          theme: ThemeData(
-            pageTransitionsTheme: const PageTransitionsTheme(
-              builders: <TargetPlatform, PageTransitionsBuilder>{
-                TargetPlatform.android: FadeUpwardsPageTransitionsBuilder(),
-              },
-            ),
-          ),
-          routes: <String, WidgetBuilder>{
-            '/next': (BuildContext context) {
-              return const Material(child: Text('Page 2'));
-            },
-          },
-        ),
-      );
-
-      final Offset widget1TopLeft = tester.getTopLeft(find.text('Page 1'));
-
-      tester.state<NavigatorState>(find.byType(Navigator)).pushNamed('/next');
-      await tester.pump();
-      await tester.pump(const Duration(milliseconds: 1));
-
-      FadeTransition widget2Opacity = tester
-          .element(find.text('Page 2'))
-          .findAncestorWidgetOfExactType<FadeTransition>()!;
-      Offset widget2TopLeft = tester.getTopLeft(find.text('Page 2'));
-      final Size widget2Size = tester.getSize(find.text('Page 2'));
-
-      // Android transition is vertical only.
-      expect(widget1TopLeft.dx == widget2TopLeft.dx, true);
-      // Page 1 is above page 2 mid-transition.
-      expect(widget1TopLeft.dy < widget2TopLeft.dy, true);
-      // Animation begins 3/4 of the way up the page.
-      expect(widget2TopLeft.dy < widget2Size.height / 4.0, true);
-      // Animation starts with page 2 being near transparent.
-      expect(widget2Opacity.opacity.value < 0.01, true);
-
-      await tester.pump(const Duration(milliseconds: 300));
-
-      // Page 2 covers page 1.
-      expect(find.text('Page 1'), findsNothing);
-      expect(find.text('Page 2'), isOnstage);
-
-      tester.state<NavigatorState>(find.byType(Navigator)).pop();
-      await tester.pump();
-      await tester.pump(const Duration(milliseconds: 1));
-
-      widget2Opacity = tester
-          .element(find.text('Page 2'))
-          .findAncestorWidgetOfExactType<FadeTransition>()!;
-      widget2TopLeft = tester.getTopLeft(find.text('Page 2'));
-
-      // Page 2 starts to move down.
-      expect(widget1TopLeft.dy < widget2TopLeft.dy, true);
-      // Page 2 starts to lose opacity.
-      expect(widget2Opacity.opacity.value < 1.0, true);
-
-      await tester.pump(const Duration(milliseconds: 300));
-
-      expect(find.text('Page 1'), isOnstage);
-      expect(find.text('Page 2'), findsNothing);
-    },
-    variant: TargetPlatformVariant.only(TargetPlatform.android),
-  );
-
-  testWidgets(
     'test page transition (CupertinoPageTransition)',
     (WidgetTester tester) async {
       final Key page2Key = UniqueKey();
