+++ packages/flutter/test/widgets/page_transitions_builder_test.dart
@@ -175,6 +175,156 @@ void main() {
     expect(find.text('Navigate'), findsNothing);
     expect(find.text('Second Page'), findsOneWidget);
   });
+
+  testWidgets('FadeUpwardsPageTransitionsBuilder test', (WidgetTester tester) async {
+    const FadeUpwardsPageTransitionsBuilder transitionsBuilder =
+        FadeUpwardsPageTransitionsBuilder();
+
+    await tester.pumpWidget(
+      Directionality(
+        textDirection: TextDirection.ltr,
+        child: Navigator(
+          onGenerateRoute: (RouteSettings settings) {
+            return _CustomPageRoute<void>(
+              settings: settings,
+              transitionsBuilder: transitionsBuilder,
+              builder: (BuildContext context) {
+                if (settings.name == '/') {
+                  return ColoredBox(
+                    color: const Color(0xFF2196F3),
+                    child: Center(
+                      child: GestureDetector(
+                        onTap: () {
+                          Navigator.of(context).pushNamed('/second');
+                        },
+                        child: const Text(
+                          'Page 1',
+                          style: TextStyle(color: Color(0xFFFFFFFF), fontSize: 24),
+                        ),
+                      ),
+                    ),
+                  );
+                }
+                return const ColoredBox(
+                  color: Color(0xFF4CAF50),
+                  child: Center(
+                    child: Text('Page 2', style: TextStyle(color: Color(0xFFFFFFFF), fontSize: 24)),
+                  ),
+                );
+              },
+            );
+          },
+        ),
+      ),
+    );
+
+    expect(find.text('Page 1'), findsOneWidget);
+    expect(find.text('Page 2'), findsNothing);
+
+    final Offset initialPage1Position = tester.getTopLeft(find.text('Page 1'));
+
+    await tester.tap(find.text('Page 1'));
+    await tester.pump();
+    await tester.pump(const Duration(milliseconds: 1));
+
+    expect(find.text('Page 1'), findsOneWidget);
+    expect(find.text('Page 2'), findsOneWidget);
+
+    final Finder page2Finder = find.text('Page 2');
+    final FadeTransition fadeTransition = tester
+        .element(page2Finder)
+        .findAncestorWidgetOfExactType<FadeTransition>()!;
+
+    // Early in the animation, opacity should be low.
+    expect(fadeTransition.opacity.value, greaterThan(0.0));
+    expect(fadeTransition.opacity.value, lessThan(0.1));
+
+    // Verify only vertical slide animation for incoming page.
+    final Offset page2Position = tester.getTopLeft(find.text('Page 2'));
+    expect(initialPage1Position.dx, equals(page2Position.dx));
+    expect(page2Position.dy, greaterThan(initialPage1Position.dy));
+
+    await tester.pumpAndSettle();
+
+    expect(find.text('Page 1'), findsNothing);
+    expect(find.text('Page 2'), findsOneWidget);
+
+    // After animation completes, verify full opacity.
+    final FadeTransition fadeTransitionComplete = tester
+        .element(find.text('Page 2'))
+        .findAncestorWidgetOfExactType<FadeTransition>()!;
+    expect(fadeTransitionComplete.opacity.value, equals(1.0));
+  });
+
+  testWidgets(
+    'FadeUpwardsPageTransitionsBuilder test with Material PageTransitionTheme',
+    (WidgetTester tester) async {
+      await tester.pumpWidget(
+        MaterialApp(
+          home: const Material(child: Text('Page 1')),
+          theme: ThemeData(
+            pageTransitionsTheme: const PageTransitionsTheme(
+              builders: <TargetPlatform, PageTransitionsBuilder>{
+                TargetPlatform.android: FadeUpwardsPageTransitionsBuilder(),
+              },
+            ),
+          ),
+          routes: <String, WidgetBuilder>{
+            '/next': (BuildContext context) {
+              return const Material(child: Text('Page 2'));
+            },
+          },
+        ),
+      );
+
+      final Offset widget1TopLeft = tester.getTopLeft(find.text('Page 1'));
+
+      tester.state<NavigatorState>(find.byType(Navigator)).pushNamed('/next');
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 1));
+
+      FadeTransition widget2Opacity = tester
+          .element(find.text('Page 2'))
+          .findAncestorWidgetOfExactType<FadeTransition>()!;
+      Offset widget2TopLeft = tester.getTopLeft(find.text('Page 2'));
+      final Size widget2Size = tester.getSize(find.text('Page 2'));
+
+      // Android transition is vertical only.
+      expect(widget1TopLeft.dx == widget2TopLeft.dx, true);
+      // Page 1 is above page 2 mid-transition.
+      expect(widget1TopLeft.dy < widget2TopLeft.dy, true);
+      // Animation begins 3/4 of the way up the page.
+      expect(widget2TopLeft.dy < widget2Size.height / 4.0, true);
+      // Animation starts with page 2 being near transparent.
+      expect(widget2Opacity.opacity.value < 0.01, true);
+
+      await tester.pump(const Duration(milliseconds: 300));
+
+      // Page 2 covers page 1.
+      expect(find.text('Page 1'), findsNothing);
+      expect(find.text('Page 2'), isOnstage);
+
+      tester.state<NavigatorState>(find.byType(Navigator)).pop();
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 1));
+
+      widget2Opacity = tester
+          .element(find.text('Page 2'))
+          .findAncestorWidgetOfExactType<FadeTransition>()!;
+      widget2TopLeft = tester.getTopLeft(find.text('Page 2'));
+
+      // Page 2 starts to move down.
+      expect(widget1TopLeft.dy < widget2TopLeft.dy, true);
+      // Page 2 starts to lose opacity.
+      expect(widget2Opacity.opacity.value < 1.0, true);
+
+      await tester.pump(const Duration(milliseconds: 300));
+
+      expect(find.text('Page 1'), isOnstage);
+      expect(find.text('Page 2'), findsNothing);
+    },
+    variant: TargetPlatformVariant.only(TargetPlatform.android),
+  );
 }
 
 class _CustomPageRoute<T> extends PageRoute<T> {
